add describe

+++ Prj01/src/com/epam/furniture/Chair.java
@@ -1,11 +1,11 @@
 package com.epam.furniture;
 
 public class Chair extends Furniture {
-    public Chair(String nameOfFurn, double square1) {
+    public Chair(String nameOfFurn, int square1) {
         super(nameOfFurn, square1);
     }
 
-    public Chair(String nameOfFurn, double square1, double square2) {
+    public Chair(String nameOfFurn, int square1, int square2) {
         super(nameOfFurn, square1, square2);
     }
 }
