+++ Prj01/src/com/epam/furniture/Sofa.java
@@ -1,11 +1,11 @@
 package com.epam.furniture;
 
 public class Sofa extends Furniture {
-    public Sofa(String nameOfFurn, double square1) {
+    public Sofa(String nameOfFurn, int square1) {
         super(nameOfFurn, square1);
     }
 
-    public Sofa(String nameOfFurn, double square1, double square2) {
+    public Sofa(String nameOfFurn, int square1, int square2) {
         super(nameOfFurn, square1, square2);
     }
 }
